docker: Stop using SSL/TLS for email
The error log was giving complaints when I tried sending invitation
emails:

> ssl.SSLCertVerificationError: [SSL: CERTIFICATE_VERIFY_FAILED] certificate verify failed: self-signed certificate (_ssl.c:1000)

Turning off TLS and SSL appears to have done the trick, and it seems a
reasonable approach since postfix is running locally (on the host, not
in the container). My emails are sending now.

Also, I added a rw mount in /tmp for easily creating files on the host
from the container.